Working multiplatform.
Removed the fullscreen feature because it screwed with Linux.
Added a script that would take the proteins and put them into one file.
Changed proteios.pde to read in the one file instead of the many files.
Binaries for all three systems in the ./binary folder.
Created the applet. For this to work, the frame settings need to be commented out in proteios.pde. However, in this commit, they are set because the exported applications can still use these settings.
Changed the README to reflect these changes.